ops: add proximity to production
Add `nginx` service to compose. Nginx proxies all incoming requests to
host machine ports, this allows proxying even if the project was run
locally and not in docker. Since we need to change addresses and ports
in the configuration add the `dev-entrypoint.sh` script.
Add also `website` service that would be the main page.